Rebasing from latest Kyigo version

+++ tg_bot/modules/cleaner.py
@@ -6,7 +6,6 @@ from tg_bot.modules.helper_funcs.chat_status import (
     bot_can_delete,
     connection_status,
     dev_plus,
-    user_admin,
 )
 from tg_bot.modules.sql import cleaner_sql as sql
 from telegram import ParseMode, Update
@@ -15,14 +14,11 @@ from telegram.ext import (
     CommandHandler,
     Filters,
     MessageHandler,
-    run_async,
 )
 
-if ALLOW_EXCL:
-    CMD_STARTERS = ("/", "!")
-else:
-    CMD_STARTERS = "/"
+from ..modules.helper_funcs.anonymous import user_admin, AdminPerms
 
+CMD_STARTERS = ("/", "!") if ALLOW_EXCL else "/"
 BLUE_TEXT_CLEAN_GROUP = 13
 CommandHandlerList = (CommandHandler, CustomCommandHandler, DisableAbleCommandHandler)
 command_list = [
@@ -50,28 +46,27 @@ def clean_blue_text_must_click(update: Update, context: CallbackContext):
     bot = context.bot
     chat = update.effective_chat
     message = update.effective_message
-    if chat.get_member(bot.id).can_delete_messages:
-        if sql.is_enabled(chat.id):
-            fst_word = message.text.strip().split(None, 1)[0]
+    if chat.get_member(bot.id).can_delete_messages and sql.is_enabled(chat.id):
+        fst_word = message.text.strip().split(None, 1)[0]
 
-            if len(fst_word) > 1 and any(
-                fst_word.startswith(start) for start in CMD_STARTERS
-            ):
+        if len(fst_word) > 1 and any(
+            fst_word.startswith(start) for start in CMD_STARTERS
+        ):
 
-                command = fst_word[1:].split("@")
-                chat = update.effective_chat
+            command = fst_word[1:].split("@")
+            chat = update.effective_chat
 
-                ignored = sql.is_command_ignored(chat.id, command[0])
-                if ignored:
-                    return
+            ignored = sql.is_command_ignored(chat.id, command[0])
+            if ignored:
+                return
 
-                if command[0] not in command_list:
-                    message.delete()
+            if command[0] not in command_list:
+                message.delete()
 
 
 @connection_status
 @bot_can_delete
-@user_admin
+@user_admin(AdminPerms.CAN_CHANGE_INFO)
 def set_blue_text_must_click(update: Update, context: CallbackContext):
     chat = update.effective_chat
     message = update.effective_message
@@ -97,17 +92,14 @@ def set_blue_text_must_click(update: Update, context: CallbackContext):
             message.reply_text(reply)
     else:
         clean_status = sql.is_enabled(chat.id)
-        if clean_status:
-            clean_status = "Enabled"
-        else:
-            clean_status = "Disabled"
+        clean_status = "Enabled" if clean_status else "Disabled"
         reply = "Bluetext cleaning for <b>{}</b> : <b>{}</b>".format(
             chat.title, clean_status
         )
         message.reply_text(reply, parse_mode=ParseMode.HTML)
 
 
-@user_admin
+@user_admin(AdminPerms.CAN_CHANGE_INFO)
 def add_bluetext_ignore(update: Update, context: CallbackContext):
     message = update.effective_message
     chat = update.effective_chat
@@ -128,7 +120,7 @@ def add_bluetext_ignore(update: Update, context: CallbackContext):
         message.reply_text(reply)
 
 
-@user_admin
+@user_admin(AdminPerms.CAN_CHANGE_INFO)
 def remove_bluetext_ignore(update: Update, context: CallbackContext):
     message = update.effective_message
     chat = update.effective_chat
@@ -151,7 +143,7 @@ def remove_bluetext_ignore(update: Update, context: CallbackContext):
         message.reply_text(reply)
 
 
-@user_admin
+@user_admin(AdminPerms.CAN_CHANGE_INFO)
 def add_bluetext_ignore_global(update: Update, context: CallbackContext):
     message = update.effective_message
     args = context.args
